cilium: Don't report health error when disabled
When the user configures --enable-health-checking=false, 'cilium status'
would previously print:

  $ cilium status
  ...
  Cluster health:               Warning   cilium-health daemon unreachable

This would occur despite the user explicitly disabling the feature. To
better reflect whether the feature is enabled or not, check in the
status response whether there is a health endpoint. If there is one,
then the feature is enabled and we can query & print its status. If
there isn't one, then the feature is disabled and we won't get a
successful response anyway, so just report back that the feature is
disabled instead:

  $ cilium status
  ...
  Cluster health:               Probe disabled

Signed-off-by: Joe Stringer <joe@cilium.io>