envoy: Register k8s service redirection to CRD listener
Register L7 LB service redirections. Use 'L7LBInfo' to keep track of
the CiliumEnvoyConfig that requested a service to be registered for L7
load balancing. Only one CiliumEnvoyConfig may request redirection of
a given service at a time, as otherwise the choice between Envoy
listeners would be ambiguous.

Service backends are synced to Envoy, so that the ingress service can
load balance to them. Datapath L7LB redirection is only requested for
'services' in a CiliumEnvoyResources. Backends are synced to Envoy
also for 'backendServices'. This allows the regular eBPF LB to be used
for pod-to-pod traffic for 'backendServices' while L7 LB is used for
ingress to those same services.

Note: 'Ingress' CEC configurations are still experimental. 

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>