datapath: Fix egress gateway to exclude in-cluster traffic
datapath code checks the egress gateway map to determine whether the packet
should be forwarded to the egress gateway and whether the packet should
be SNAT-ed with an appropriate egress IP.

In-cluster traffic should be, however, excluded from these checks.
In-cluster traffic includes both pod-to-pod and pod-to-node. The purpose
of this patch is to distinguish between cluster-local versus
clluster-egress traffic, and apply the egress map check only in the
latter case.

There are two points we do so:
 - in bpf_lxc where we check whether the packet should be forwarded to
 the egress gateway (currently, via a tunnel)
 - in bpf_host (in the gateway node) where we check whether the packet should
 be SNATed

The checks in these two points are as follows:

In bpf_lxc, we check whether:
 - there is tunnel_endpoint in ipcache, which means the destination is a
 remote endpoint
 - the destination id is REMOTE_NODE_ID or HOST_ID, which means that the
 destination is the host or a remote node
 - the destination matches a local endpoint
and in all above cases we skip checking the egress policy map.

In bpf_host, we only perform the egress map check if:
 - the source address does not match a local endpoint
 - the remote address is not a remote node

One of the things that we considered was using WORLD_ID to distinguish
egress traffic. However, as Paul pointed out, egress traffic might
not have a WORLD_ID because it might match CIDR policies (toCIDR rules).

Fix issue:16147

Signee-off-by: Yongkun Gui <ygui@google.com>
Signed-off-by: Martynas Pumputis <m@lambda.lt>
Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>